db(minutes): apply both ledger migrations live, and cast the roster sum back to bigint
The class-roster migration would not apply as written: play_seconds is bigint,
sum(bigint) is numeric, and CREATE OR REPLACE VIEW may not change a column's
type — 'cannot change data type of view column total_practice_seconds from
bigint to numeric'. The cast is lossless, play_seconds holding whole seconds,
and keeps the column exactly the bigint it has always been. The migration's own
test was written against the file's text rather than the live database, so it
was green either way; this is what running it found.

BOTH MIGRATIONS ARE NOW APPLIED TO THE LIVE DATABASE, each in one transaction,
each ending in NOTIFY pgrst.

Spot-checked live, admin RPC then class roster then the learner's own figure:

  Nina Chopra, fra_for_eng, class 8A French
    roster 173 min -> 160, admin 173 -> 160, learner reads 160. Exact.
  Stephen   admin 43,958 min -> 5,408 against a learner-facing 5,320
  Tom       admin 26,493 min -> 2,810 against a learner-facing 2,780

The residual on those last two is the position-derived estimate branch, firing
for courses they enrolled on and never played — 15 such enrolments between them
— and the rows come back flagged is_estimated, exactly as the migration says.
Four further roster rows checked: all now equal to the ledger to the minute.

schema.sql regenerated per supabase/migrations/README.md. It also picks up
handbook_questions from job #386, which was applied live today and not yet
snapshotted; that is the snapshot catching up, not a change made here.

+++ supabase/migrations/20260908d_class_progress_from_playback_ledger.sql
@@ -37,9 +37,13 @@ SELECT c.id AS class_id,
     COALESCE(( SELECT count(*) AS count
            FROM lego_progress lp
           WHERE lp.learner_id = l.id AND lp.course_id = c.course_code AND lp.is_retired = true), 0::bigint) AS legos_mastered,
+    -- play_seconds is bigint and sum(bigint) is numeric, so the cast is what
+    -- keeps this column bigint, as it has always been. Lossless: play_seconds
+    -- holds whole seconds. Without it CREATE OR REPLACE VIEW refuses, because
+    -- it may not change a column's type.
     COALESCE(( SELECT sum(lso.play_seconds) AS sum
            FROM learner_speaking_opportunities lso
-          WHERE lso.learner_id = l.id AND lso.course_code = c.course_code), 0::bigint) AS total_practice_seconds,
+          WHERE lso.learner_id = l.id AND lso.course_code = c.course_code), 0::numeric)::bigint AS total_practice_seconds,
     ( SELECT max(s.ended_at) AS max
            FROM sessions s
           WHERE s.learner_id = l.id AND s.course_id = c.course_code) AS last_active_at,
